legacy/Archive/pdfSummaryGPT_function-source (1)/requirements.txt
Flask~=2.2.5
openai==0.27.0
google-api-python-client==2.93.0
google-auth==2.22.0
google-auth-oauthlib==1.0.0
google-auth-httplib2==0.1.0
google-cloud-storage==2.9.0
google-cloud-core==2.3.3
google-cloud-vision==3.3.1
PyPDF2==3.0.1
python-docx==0.8.11
requests==2.32.0
pytz==2024.2
pdf2image==1.17.0
pytesseract==0.3.13
protobuf==4.24.3
google-generativeai==0.3.0